EnhancedStorage improvements and browser-specific screenshots.
