+++ src/routes/embedded/support/unpartitioned-state/unpartitioned-state.view.tsx
@@ -1,20 +1,42 @@
 import { useEmbedded } from "~utils/embedded/embedded.hooks";
-import { useCallback, useEffect, useState } from "react";
+import React, { useCallback, useEffect, useState } from "react";
 import { useLocation } from "~wallets/router/router.utils";
 import { OnboardingCard } from "~components/embed/ui/molecules/card/onboarding-card/OnboardingCard";
 import { EmbeddedPaths } from "~wallets/router/iframe/iframe.routes";
 import { Button, Checkbox } from "~components/embed";
 import { LocalStorage } from "~iframe/storage/unpartitioned-storage/local-storage";
+import Image from "~components/common/Image";
+import { toast } from "react-toastify";
+import { FormattedText } from "~components/embed/ui/atoms/formatted-text/FormattedText";
+
 import chromeLogoSrc from "url:assets/icons/browsers/chrome-logo.png";
 import edgeLogoSrc from "url:assets/icons/browsers/edge-logo.png";
 import operaLogoSrc from "url:assets/icons/browsers/opera-logo.png";
 import braveLogoSrc from "url:assets/icons/browsers/brave-logo.png";
 import safariLogoSrc from "url:assets/icons/browsers/safari-logo.png";
 import firefoxLogoSrc from "url:assets/icons/browsers/firefox-logo.png";
-import Image from "~components/common/Image";
+import braveScreenshotSrc from "url:assets/screenshots/unpartitioned-state/raw-brave-dark.png";
+import chrome1ScreenshotSrc from "url:assets/screenshots/unpartitioned-state/raw-chrome-1-dark.png";
+import chrome2ScreenshotSrc from "url:assets/screenshots/unpartitioned-state/raw-chrome-2-dark.png";
+import chrome3ScreenshotSrc from "url:assets/screenshots/unpartitioned-state/raw-chrome-3-dark.png";
+import inAppScreenshotSrc from "url:assets/screenshots/unpartitioned-state/raw-in-app-dark.png";
 
 import styles from "./unpartitioned-state.module.scss";
-import { toast } from "react-toastify";
+
+const pretendToBeBrave = false;
+const isBrave = pretendToBeBrave || window.navigator.brave;
+
+const pretendToBeMobileChrome = false;
+const isMobileChrome = pretendToBeMobileChrome || false;
+
+const pretendToBeInAppAndroidBrowser = false;
+const isInAppAndroidBrowser =
+  pretendToBeInAppAndroidBrowser ||
+  (navigator.userAgent.includes("Android") &&
+    (navigator.userAgent.includes("wv") || navigator.userAgent.includes("WebView")));
+
+// Note this flags are only useful to see the UI in those browsers, but not to reproduce their exact behavior.
+const isPretendingToBeAnotherBrowser = pretendToBeBrave || pretendToBeMobileChrome || pretendToBeInAppAndroidBrowser;
 
 export function UnpartitionedStateMissingEmbeddedView() {
   const { navigate } = useLocation();
@@ -50,8 +72,10 @@ export function UnpartitionedStateMissingEmbeddedView() {
 
   const handleRequestPermission = useCallback(async () => {
     try {
-      await LocalStorage.getInstance();
-      console.log("COULD CONTINUE");
+      const localStorage = await LocalStorage.getInstance();
+
+      console.log("COULD CONTINUE", localStorage.status);
+
       navigate(EmbeddedPaths.Auth);
     } catch (error) {
       console.log(error);
@@ -80,10 +104,6 @@ export function UnpartitionedStateMissingEmbeddedView() {
     "Your browser doesn't support cross-site authentication and wallet syncing. You'll need to manually import your wallet on each new site.";
 
   // Logos downloaded from https://github.com/alrra/browser-logos
-
-  // TODO: Consider showing different browsers in mobile than in desktop?
-
-  // TODO: Consider adding Android/Samsung browser?
 
   let children = (
     <>
@@ -146,9 +166,9 @@ export function UnpartitionedStateMissingEmbeddedView() {
 
   if (
     errorsWhileRequestingAccess < 3 &&
-    (unpartitionedStateStatus === "rejected" || unpartitionedStateStatus === "error")
+    (unpartitionedStateStatus === "rejected" || unpartitionedStateStatus === "error" || isPretendingToBeAnotherBrowser)
   ) {
-    if (unpartitionedStateStatus === "rejected") {
+    if (errorsWhileRequestingAccess === 0 || unpartitionedStateStatus === "rejected") {
       headerText = "Enable browser storage support";
     } else {
       headerText = "Error accessing browser storage";
@@ -157,10 +177,63 @@ export function UnpartitionedStateMissingEmbeddedView() {
     subtitle =
       "Before you continue, Wander Connect needs access to your browser's storage to enable cross-site authentication and wallet synching.";
 
+    let browserSpecificInstructions: React.ReactNode = null;
+
+    if (isBrave) {
+      browserSpecificInstructions = (
+        <FormattedText
+          children={[
+            <p key="text">
+              You can enable this from the <em className={styles.inlineQuote}>Embedded content</em> option in the
+              navigation bar:
+            </p>,
+            <p key="image">
+              <Image src={braveScreenshotSrc} borderRadius={8} />
+            </p>,
+          ]}
+        />
+      );
+    } else if (isMobileChrome) {
+      browserSpecificInstructions = (
+        <FormattedText
+          children={[
+            <p key="text">
+              You can enable this from the <em className={styles.inlineQuote}>Cookies and site data</em> option in the
+              navigation bar:
+            </p>,
+            <p key="image1">
+              <Image src={chrome1ScreenshotSrc} borderRadius={8} />
+            </p>,
+            <p key="image2">
+              <Image src={chrome2ScreenshotSrc} borderRadius={8} />
+            </p>,
+            <p key="image3">
+              <Image src={chrome3ScreenshotSrc} borderRadius={8} />
+            </p>,
+          ]}
+        />
+      );
+    } else if (isInAppAndroidBrowser) {
+      browserSpecificInstructions = (
+        <FormattedText
+          children={[
+            <p key="text">You'll have to switch to your browser app to enable this:</p>,
+            <p key="image">
+              <Image src={inAppScreenshotSrc} borderRadius={8} />
+            </p>,
+          ]}
+        />
+      );
+    }
+
     children = (
-      <Button variant="primary" size="md" isDisabled={areButtonsDisabled} onClick={handleRequestPermission}>
-        {unpartitionedStateStatus === "rejected" ? "Request access" : "Try again"}
-      </Button>
+      <>
+        {browserSpecificInstructions}
+
+        <Button variant="primary" size="md" isDisabled={areButtonsDisabled} onClick={handleRequestPermission}>
+          {unpartitionedStateStatus === "rejected" ? "Request access" : "Try again"}
+        </Button>
+      </>
     );
   }
